Introduce GitHub Actions script for automated packaging
This commit add a GitHub Actions script that enables automated
pre-release testing, tagging, and periodical updating of the PSL.
When a new commit on the master branch is verified by GitHub Actions,
it will be tagged as `vX.Y.Z`.
Additionaly, it will be released as a new package with a version number
like `X.Y.Z.YYYYmmdd` for each PSL update.
In this commit, autorelease script still uses TestPyPI for testing.