logic: Bulk rewrite to adapt the latest PSL rules
Rewriting most of evaluation code to comply with the latest PSL
definitions.

The current PSL definition has conflicting evaluation rules with the old
test data in test_psl.txt.  We cannot satisfy the both. This module will
follow the PSL wiki.

With this rewrite, the API functions start to accept tuple of bytes as
input domain name.
This changge is required to support arbitrary bytes, including dots or
null as the input.

Note that this module do not check the correctness of the bytes. When
one need to convert the returned tuple of bytes to str to display, the
caller must check the returned value is apropriate encoding, escaping,
etc.

Signed-off-by: ko-zu <causeless@gmail.com>