logic: follow the PSL linter evaluation rules
This commit reverts that internal logic change.

In the previous commit 72111bdc835ba this module changed the evaluation
logic to follow the PSL definition described in the wiki.

Now, it follows the rule described in the linter of the PSL code
repository, where a wildcard declaration implies that its zone root is
also a public suffix.

Signed-off-by: ko-zu <causeless@gmail.com>