ci: do not check PSL freshness when releasing master branch
Remove the verification logic that ensures the master branch has the
up-to-date PSL when releasing a new version.

It seems the PSL publishing point is distributing different versions of
the PSL for hours. This causes the automation logic to fail randomly. We
cannot rely on its consistency.

In the prior release github actions get:
last-modified: Fri, 21 Jun 2024 13:05:36 GMT
https://github.com/ko-zu/psl/actions/runs/9616800387/job/26527141543

Fetching the same address from a devel machine:
last-modified: Wed, 19 Jun 2024 11:00:02 GMT

Signed-off-by: ko-zu <causeless@gmail.com>